requirements.txt
pytest==8.0.0
requests==2.31.0
pytest-asyncio==0.23.5
aiohttp==3.9.3
pytest-cov==4.1.0
python-dotenv==1.0.0
tenacity==8.2.3
structlog==24.1.0